app/FastApi/requirements.txt
fastapi==0.121.2
uvicorn==0.38.0
python-dotenv==1.2.1
transformers==4.57.3
onnxruntime==1.23.2
numpy==2.2.6
azure-monitor-opentelemetry==1.8.2
